products/requirements.txt
asgiref==3.10.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
cryptography==46.0.2
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==4.2.25
django-admin-interface==0.31.0
django-allauth==65.12.0
django-cache-url==3.4.5
django-colorfield==0.14.0
django-crispy-forms==2.4
django-grappelli==4.0.2
django-jalali-date==2.0.0
django-multi-captcha-admin==2.0.0
django-recaptcha2==1.4.1
django-rosetta==0.10.2
environs==14.4.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
marshmallow==4.0.1
oauthlib==3.3.1
pillow==11.3.0
polib==1.2.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
python-gettext==5.0
python-openid==2.2.5
python-slugify==8.0.4
requests==2.32.5
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2021.5
urllib3==2.5.0
